requirements.txt
asgiref==3.4.1
attrs==21.2.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
git+https://github.com/andresionek91/CorreiosPrecoPrazo
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
model-mommy==2.0.0
packaging==21.2
Pillow==8.4.0
pycodestyle==2.8.0
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
xmltodict==0.12.0
zipp==3.6.0
gunicorn==20.0.4